Update for issue 20
If a line begins with "%" in the date file, it will be ignored by the
parser.

Cases which the current changes in the parser will not be able to handle
when the information regarding a bus is commented :
1. The bus is being used in some branch definition.
2. The generator information regarding the bus is still present in the
file.
3. The bus location regarding the bus is still present.

All the above cases create an anomaly in the file. Support for the above
will be added if such cases are encountered.